Prep for compiler driver
