+++ midlang/src/errors.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone)]
+pub struct ParseError;
+
+#[derive(Debug, Clone)]
+pub struct BuildError;
